+++ src/main/java/de/mossgrabers/convertwithmoss/core/ConverterBackend.java
@@ -51,6 +51,8 @@ import de.mossgrabers.convertwithmoss.format.elektron.TonverkMultiCreator;
 import de.mossgrabers.convertwithmoss.format.elektron.TonverkMultiDetector;
 import de.mossgrabers.convertwithmoss.format.elektron.TonverkPresetCreator;
 import de.mossgrabers.convertwithmoss.format.elektron.TonverkPresetDetector;
+import de.mossgrabers.convertwithmoss.format.emu.emulator4.Emulator4Creator;
+import de.mossgrabers.convertwithmoss.format.emu.emulator4.Emulator4Detector;
 import de.mossgrabers.convertwithmoss.format.ensoniq.epsasr.EnsoniqEpsAsrDetector;
 import de.mossgrabers.convertwithmoss.format.ensoniq.mirage.MirageDetector;
 import de.mossgrabers.convertwithmoss.format.exs.EXS24Creator;
@@ -159,6 +161,7 @@ public class ConverterBackend
         this.detectors.add (new DecentSamplerDetector (notifier));
         this.detectors.add (new DlsDetector (notifier));
         this.detectors.add (new DistingExDetector (notifier));
+        this.detectors.add (new Emulator4Detector (notifier));
         this.detectors.add (new TonverkMultiDetector (notifier));
         this.detectors.add (new TonverkPresetDetector (notifier));
         this.detectors.add (new EnsoniqEpsAsrDetector (notifier));
@@ -199,6 +202,7 @@ public class ConverterBackend
         this.creators.add (new TX16WxCreator (notifier));
         this.creators.add (new DecentSamplerCreator (notifier));
         this.creators.add (new DistingExCreator (notifier));
+        this.creators.add (new Emulator4Creator (notifier));
         this.creators.add (new TonverkMultiCreator (notifier));
         this.creators.add (new TonverkPresetCreator (notifier));
         this.creators.add (new KMPCreator (notifier));